web/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
mysqlclient
python-dotenv
fastapi-admin
aiofiles
python-jose[cryptography]
passlib[bcrypt]
